+++ core/src/main/kotlin/no/elg/hex/Hex.kt
@@ -8,6 +8,7 @@ import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.Input.Keys
 import com.badlogic.gdx.InputMultiplexer
 import com.badlogic.gdx.Preferences
+import com.badlogic.gdx.audio.Music
 import com.badlogic.gdx.graphics.Color
 import com.badlogic.gdx.graphics.Cursor
 import com.badlogic.gdx.graphics.GL20
@@ -21,11 +22,14 @@ import com.fasterxml.jackson.dataformat.smile.databind.SmileMapper
 import com.fasterxml.jackson.module.kotlin.jsonMapper
 import com.fasterxml.jackson.module.kotlin.kotlinModule
 import com.kotcrab.vis.ui.VisUI
+import ktx.assets.disposeSafely
 import ktx.async.AsyncExecutorDispatcher
 import ktx.async.KtxAsync
 import ktx.async.newSingleThreadAsyncContext
 import no.elg.hex.Settings.MSAA_SAMPLES_PATH
+import no.elg.hex.audio.DisabledMusicHandler
 import no.elg.hex.audio.MusicHandler
+import no.elg.hex.audio.RealMusicHandler
 import no.elg.hex.event.Events
 import no.elg.hex.hud.GLProfilerRenderer
 import no.elg.hex.hud.MessagesRenderer
@@ -105,7 +109,8 @@ object Hex : ApplicationAdapter() {
   lateinit var platform: Platform
 
   lateinit var launchPreference: Preferences
-  var audioDisabled: Boolean = true
+
+  val music: MusicHandler get() = MusicHandler.instance ?: error("Music handler not setup!")
 
   val debugStage by lazy { trace || args.`stage-debug` }
   val debug by lazy { args.debug || args.trace }
@@ -142,8 +147,6 @@ object Hex : ApplicationAdapter() {
       }
       Gdx.graphics.requestRendering()
     }
-
-  val music: MusicHandler = MusicHandler()
 
   override fun create() {
     if (args.`reset-all`) {